+++ src/sbml/packages/fbc/sbml/FbcAnd.cpp
@@ -275,7 +275,7 @@ FbcAnd::toInfix() const
   for (size_t pos = 1; pos < mAssociations.size(); ++pos)
   {
     str << " and ";
-    str << mAssociations.get(pos)->toInfix();
+    str << mAssociations.get((unsigned int)pos)->toInfix();
   }
   str << ")";
   return str.str();
@@ -605,20 +605,20 @@ FbcAnd::readAttributes (const XMLAttributes& attributes,
   if (getErrorLog() != NULL)
   {
     numErrs = getErrorLog()->getNumErrors();
-    for (int n = numErrs-1; n >= 0; n--)
+    for (int n = (int)numErrs-1; n >= 0; n--)
     {
-      if (getErrorLog()->getError(n)->getErrorId() == UnknownPackageAttribute)
+      if (getErrorLog()->getError((unsigned int)n)->getErrorId() == UnknownPackageAttribute)
       {
         const std::string details =
-                          getErrorLog()->getError(n)->getMessage();
+                          getErrorLog()->getError((unsigned int)n)->getMessage();
         getErrorLog()->remove(UnknownPackageAttribute);
         getErrorLog()->logPackageError("fbc", FbcUnknown,
                        getPackageVersion(), sbmlLevel, sbmlVersion, details, getLine(), getColumn());
       }
-      else if (getErrorLog()->getError(n)->getErrorId() == UnknownCoreAttribute)
+      else if (getErrorLog()->getError((unsigned int)n)->getErrorId() == UnknownCoreAttribute)
       {
         const std::string details =
-                          getErrorLog()->getError(n)->getMessage();
+                          getErrorLog()->getError((unsigned int)n)->getMessage();
         getErrorLog()->remove(UnknownCoreAttribute);
         getErrorLog()->logPackageError("fbc", FbcUnknown,
                        getPackageVersion(), sbmlLevel, sbmlVersion, details, getLine(), getColumn());
